+++ CHANGES.md
@@ -8,6 +8,7 @@
 
 ### Changes
 * The old functionality to upload one photo or video at a time to the HAP+ server can be enabled in the main iOS settings app
+* Displayed the name of the file being uploaded on the "file exists" alert, so that it is known which file the message is referring to
 
 ## 0.7.0